Add the initial piano roll and fix jack startup.
Add the initial piano roll code.  This is very early code, it has mainly been
added to get it out of my client.

Fix jack startup so that we start the WaveManager before fluidsynth.  This
seems to be the only way to get everything to work, if we start the
WaveManager afterwards fluidsynth's jack processing thread seems to get
starved.  This is likely related to jitting, I haven't seen it occur on an AOT
compiled binary.