Add safety check in reposition after record
If we loop around while repositioning a track cursor after record, consider
that the new position.

This situation shouldn't happen, because we always evaluate time mod track
span and the last event in the track should always be at t = track span.
Likewise, we don't see an error for it.