Make wavtool mix using the same audio processing code as the workbench.
wavtool previously had its own audio processing code, which was intended to
mirror that of the jack module and processing modules.

Add the render module, which implements the wavtool crash command set on a
rendering engine that runs through JackWrapper in a "dumb" mode so there's no
actual jack connection.  This should ensure that a randered wav file is
exactly what you heard when mixing with jack.