Add initial implementation of a "TimeBar"
Add the "TimeBar" which will eventually be a UI control that shows our
location in the section graphically as well as numerically.

This is currently pretty weak (it only shows time in millis since start, and
since forward and backward are still the old event manager versions they don't
work correctly with it) but it introduces an important architectural
change, as textui now has its own Poller and provides a poll queue to allow
other threads to send messages to be executed in the main, UI thread.