Fix crash context processing in textui.
Always copy the current context when creating a CrashCallback so we get the
context we're initiated in and can't mess up a context that has moved on.
Also define RunUI as a class so we can grab the current context instead of the
root context (this doesn't matter as much but it's preferrable because we
can't be sure that the root context is necessarily the same at the point when
we're called).