Remove local device configuration from awbconfig and allow setting midi output.
Remove all of the local device configuration (specific to devices that I
happen to have on one of my machines) and make it so everything (specifically
the "track controller functions") can be defined from .awb.crash.

Also add a button to the UI to allow you to set the midi output channel.