Add support for recording from a position other than the origin.
Allow recording starting at a non-zero position and keep track of the position
so the wave can be played back from the correct start point.

Tracks of this type are stored as crash files that build a SequencerWaveModel
that positions the recorded wave according to the current frame position of
the JackWrapper.

There's still a few parts missing here: we probably don't write information on
these kinds of tracks to the project file correctly.  The entry we write to
the logfile for them won't work with wavtool.