Removed the "snap to measure" while storing the track ending.
This was broken and not very useful given my current goals.  Instead, adjust
the span so that it is long enough to encompass tracks that are longer than
the span (padding out by measures) and adjust the track time (time of the
EndTrack event) to the time of the span so that LoopCursor's work correctly.